feat(adaptServiceTrackToCarService.ts): add adapter function to convert ServiceTrack to CarServiceTrack
The `adaptServiceTrackToCarService` function is added to the `out` directory of the `adapters` module. This function takes a `ServiceTrack` object as input and returns a `CarServiceTrack` object. The function extracts the necessary properties from the `serviceTrack` object and maps the `services` array to a new array with renamed properties. This adapter function is used to convert a `ServiceTrack` object to a `CarServiceTrack` object in the application.